+++ src/NetTopologySuite.IO.GPX/GpxWriter.cs
@@ -76,19 +76,19 @@ namespace NetTopologySuite.IO
         /// <list type="table">
         /// <listheader>
         /// <term>Type of <see cref="IFeature.Geometry"/></term>
-        /// <term>Corresponding top-level data object</term>
+        /// <description>Corresponding top-level data object</description>
         /// </listheader>
         /// <item>
         /// <term><see cref="Point"/></term>
-        /// <term><see cref="GpxWaypoint"/></term>
+        /// <description><see cref="GpxWaypoint"/></description>
         /// </item>
         /// <item>
         /// <term><see cref="LineString"/></term>
-        /// <term><see cref="GpxRoute"/></term>
+        /// <description><see cref="GpxRoute"/></description>
         /// </item>
         /// <item>
         /// <term><see cref="MultiLineString"/></term>
-        /// <term><see cref="GpxTrack"/></term>
+        /// <description><see cref="GpxTrack"/></description>
         /// </item>
         /// </list>
         /// </remarks>
@@ -202,6 +202,12 @@ namespace NetTopologySuite.IO
             writer.WriteGpxStartElement("gpx");
             writer.WriteAttributeString("version", "1.1");
             writer.WriteAttributeString("creator", metadata.Creator);
+
+            foreach (var (desiredPrefix, namespaceUri) in settings.CommonXmlNamespacesByDesiredPrefix)
+            {
+                writer.WriteAttributeString("xmlns", desiredPrefix, null, namespaceUri.ToString());
+            }
+
             if (!metadata.IsTrivial)
             {
                 writer.WriteGpxStartElement("metadata");