+++ src/NetTopologySuite.IO.GPX/GpxWriterSettings.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Xml;
 
 namespace NetTopologySuite.IO
 {
@@ -35,5 +37,18 @@ namespace NetTopologySuite.IO
         /// is an instance of the base class (see its summary documentation for details).
         /// </summary>
         public GpxExtensionWriter ExtensionWriter { get; set; } = DefaultExtensionWriter;
+
+        /// <summary>
+        /// Gets an <see cref="IDictionary{TKey, TValue}"/> instance that can be used to register
+        /// additional namespaces to declare on the root element, keyed by the desired prefix to use
+        /// in the result.  These can be used to avoid duplicating the namespace declaration for
+        /// every single "extensions" element that uses the exact same namespace.
+        /// <para>
+        /// The <see cref="XmlWriterSettings"/>'s <see cref="XmlWriterSettings.NamespaceHandling"/>
+        /// value will need to be set to <see cref="NamespaceHandling.OmitDuplicates"/> when the
+        /// writer is created in order to fully benefit from using this property.
+        /// </para>
+        /// </summary>
+        public IDictionary<string, Uri> CommonXmlNamespacesByDesiredPrefix { get; } = new Dictionary<string, Uri>();
     }
 }